requirements.txt
wheel
discord.py
Quart

